Minor fixes

+++ src/Components/Section_Elements/Cases/CaseThumbnailLatest.js
@@ -35,7 +35,7 @@ const StyledBackgroundColor = styled(motion.div)`
   height: 100%;
   position: relative;
   overflow: hidden;
-  border-radius: ${padding.insideElements.half};
+  border-radius: ${padding.insideElements.single};
   transition: all 0.3s;
 
   ${({ $bgColor }) =>
